# Conflicts:
#	build.gradle
#	build_and_publish.sh

+++ src/main/java/net/geant/nmaas/portal/api/market/DomainController.java
@@ -1,12 +1,18 @@
 package net.geant.nmaas.portal.api.market;
 
+import lombok.RequiredArgsConstructor;
 import net.geant.nmaas.dcn.deployment.DcnDeploymentStateChangeEvent;
 import net.geant.nmaas.dcn.deployment.entities.CustomerNetwork;
 import net.geant.nmaas.dcn.deployment.entities.DcnDeploymentState;
 import net.geant.nmaas.orchestration.events.dcn.DcnDeployedEvent;
 import net.geant.nmaas.orchestration.events.dcn.DcnRemoveActionEvent;
 import net.geant.nmaas.orchestration.exceptions.InvalidDomainException;
-import net.geant.nmaas.portal.api.domain.*;
+import net.geant.nmaas.portal.api.domain.DomainBase;
+import net.geant.nmaas.portal.api.domain.DomainBaseWithState;
+import net.geant.nmaas.portal.api.domain.DomainGroupView;
+import net.geant.nmaas.portal.api.domain.DomainRequest;
+import net.geant.nmaas.portal.api.domain.DomainView;
+import net.geant.nmaas.portal.api.domain.Id;
 import net.geant.nmaas.portal.api.exception.MissingElementException;
 import net.geant.nmaas.portal.api.exception.ProcessingException;
 import net.geant.nmaas.portal.exceptions.ObjectNotFoundException;
@@ -15,10 +21,9 @@ import net.geant.nmaas.portal.persistent.entity.Domain;
 import net.geant.nmaas.portal.persistent.entity.Role;
 import net.geant.nmaas.portal.persistent.entity.User;
 import net.geant.nmaas.portal.service.ApplicationStatePerDomainService;
+import net.geant.nmaas.portal.service.DomainGroupService;
 import net.geant.nmaas.portal.service.DomainService;
-import net.geant.nmaas.portal.service.UserService;
 import org.apache.commons.lang3.StringUtils;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.transaction.annotation.Transactional;
@@ -41,32 +46,27 @@ import java.util.stream.Collectors;
 import static com.google.common.base.Preconditions.checkArgument;
 
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/api/domains")
 public class DomainController extends AppBaseController {
-
-	private final DomainService domainService;
-
-	private final ApplicationEventPublisher eventPublisher;
-
-	private final ApplicationStatePerDomainService applicationStatePerDomainService;
 
 	private static final String UNABLE_TO_CHANGE_DOMAIN_ID = "Unable to change domain id";
 	private static final String DOMAIN_NOT_FOUND = "Domain not found.";
 
-	@Autowired
-	public DomainController(UserService userService, DomainService domainService, ApplicationEventPublisher eventPublisher, ApplicationStatePerDomainService applicationStatePerDomainService){
-		this.userService = userService;
-		this.domainService = domainService;
-		this.eventPublisher = eventPublisher;
-		this.applicationStatePerDomainService = applicationStatePerDomainService;
-	}
+	private final DomainService domainService;
+	private final DomainGroupService domainGroupService;
+	private final ApplicationEventPublisher eventPublisher;
+	private final ApplicationStatePerDomainService applicationStatePerDomainService;
 
 	@GetMapping
 	@Transactional(readOnly = true)
 	@PreAuthorize("hasRole('ROLE_SYSTEM_ADMIN')")
 	public List<DomainView> getDomains() {
 		return domainService.getDomains().stream()
-				.map(d -> modelMapper.map(d, DomainView.class))
+				.map(d -> {
+					d = domainService.getAppStatesFromGroups(d);
+					return modelMapper.map(d, DomainView.class);
+				})
 				.collect(Collectors.toList());
 	}
 
@@ -77,7 +77,11 @@ public class DomainController extends AppBaseController {
 		User user = userService.findByUsername(principal.getName()).orElseThrow(() -> new ProcessingException("User not found."));
 		Domain domain = domainService.findDomain(domainId).orElseThrow(() -> new MissingElementException(DOMAIN_NOT_FOUND));
 		// if is system admin or domain admin than return full view
-		if(user.getRoles().stream().anyMatch(role -> role.getRole() == Role.ROLE_SYSTEM_ADMIN)
+
+		// check groups status of app
+		domain = domainService.getAppStatesFromGroups(domain);
+
+		if (user.getRoles().stream().anyMatch(role -> role.getRole() == Role.ROLE_SYSTEM_ADMIN)
 				|| user.getRoles().stream().anyMatch(role -> role.getDomain().getId().equals(domainId)
 				&& role.getRole() == Role.ROLE_DOMAIN_ADMIN)) {
 
@@ -102,7 +106,7 @@ public class DomainController extends AppBaseController {
 	@Transactional
 	@PreAuthorize("hasRole('ROLE_SYSTEM_ADMIN')")
 	public Id createDomain(@RequestBody(required=true) DomainRequest domainRequest) {
-		if(domainService.existsDomain(domainRequest.getName())) {
+		if (domainService.existsDomain(domainRequest.getName())) {
 			throw new ProcessingException("Domain already exists.");
 		}
 
@@ -110,7 +114,7 @@ public class DomainController extends AppBaseController {
 			Domain domain = domainService.createDomain(domainRequest);
 			this.domainService.storeDcnInfo(domain.getCodename(), domain.getDomainDcnDetails().getDcnDeploymentType());
 
-			if(domain.getDomainDcnDetails().isDcnConfigured()){
+			if (domain.getDomainDcnDetails().isDcnConfigured()) {
 				this.eventPublisher.publishEvent(new DcnDeploymentStateChangeEvent(this, domain.getCodename(), DcnDeploymentState.DEPLOYED));
 				this.eventPublisher.publishEvent(new DcnDeployedEvent(this, domain.getCodename()));
 			}
@@ -139,7 +143,7 @@ public class DomainController extends AppBaseController {
 		domain.getDomainDcnDetails().setDcnDeploymentType(domainUpdate.getDomainDcnDetails().getDcnDeploymentType());
 		domain.getDomainDcnDetails().getCustomerNetworks().clear();
 		domainUpdate.getDomainDcnDetails().getCustomerNetworks().stream().map(CustomerNetwork::of).forEach(net -> domain.getDomainDcnDetails().getCustomerNetworks().add(net));
-		if(StringUtils.isEmpty(domainUpdate.getDomainTechDetails().getExternalServiceDomain())){
+		if (StringUtils.isEmpty(domainUpdate.getDomainTechDetails().getExternalServiceDomain())) {
 			domain.getDomainTechDetails().setExternalServiceDomain(domainUpdate.getDomainTechDetails().getExternalServiceDomain());
 		} else if(!domainUpdate.getDomainTechDetails().getExternalServiceDomain().equalsIgnoreCase(domain.getDomainTechDetails().getExternalServiceDomain())){
 			checkArgument(!domainService.existsDomainByExternalServiceDomain(domainUpdate.getDomainTechDetails().getExternalServiceDomain()), "External service domain is not unique");
@@ -199,10 +203,78 @@ public class DomainController extends AppBaseController {
 	@DeleteMapping("/{domainId}")
 	@Transactional
 	@PreAuthorize("hasRole('ROLE_SYSTEM_ADMIN')")
-	public void deleteDomain(@PathVariable Long domainId) {		
-		if(!domainService.removeDomain(domainId)) {
-			throw new MissingElementException("Unable to delete domain");
+	public void deleteDomain(@PathVariable Long domainId, @RequestParam(required = false, name = "softRemove") Boolean softRemove) {
+		if (softRemove != null && softRemove) {
+			if(!domainService.softRemoveDomain(domainId)) {
+				throw new MissingElementException("Unable to soft remove domain");
+			}
+			return;
 		}
+		if(!domainService.removeDomain(domainId)) {
+			throw new MissingElementException("Unable to remove domain");
+		}
+	}
+
+	@PostMapping("/group")
+	@Transactional
+	@PreAuthorize("hasRole('ROLE_SYSTEM_ADMIN')")
+	public Id createDomainGroup(@RequestBody(required=true) DomainGroupView domainGroup) {
+		if (domainGroupService.existDomainGroup(domainGroup.getName(), domainGroup.getCodename())) {
+			throw new ProcessingException("Domain group already exists.");
+		}
+		try {
+			DomainGroupView domainGroupView = domainGroupService.createDomainGroup(domainGroup);
+			return new Id(domainGroupView.getId());
+		} catch (InvalidDomainException e) {
+			throw new ProcessingException(e.getMessage());
+		}
+	}
+
+	@DeleteMapping("/group/{domainGroupId}")
+	@Transactional
+	@PreAuthorize("hasRole('ROLE_SYSTEM_ADMIN')")
+	public void deleteDomainGroup(@PathVariable Long domainGroupId){
+		this.domainGroupService.deleteDomainGroup(domainGroupId);
+	}
+
+	@GetMapping("/group")
+	@Transactional(readOnly = true)
+	@PreAuthorize("hasRole('ROLE_SYSTEM_ADMIN')")
+	public List<DomainGroupView> getDomainGroups() {
+		return domainGroupService.getAllDomainGroups();
+	}
+
+	@GetMapping("/group/{domainGroupId}")
+	@Transactional(readOnly = true)
+	@PreAuthorize("hasRole('ROLE_SYSTEM_ADMIN')")
+	public DomainGroupView getDomainGroup(@PathVariable Long domainGroupId) {
+		return domainGroupService.getDomainGroup(domainGroupId);
+	}
+
+	@PostMapping("/group/{domainGroupCodeName}")
+	@Transactional
+	@PreAuthorize("hasRole('ROLE_SYSTEM_ADMIN')")
+	public DomainGroupView addDomainsToGroup(@PathVariable String domainGroupCodeName,
+											 @RequestBody List<Long> domainIds) {
+		return domainGroupService.addDomainsToGroup(
+				domainService.getDomains().stream().filter(d -> domainIds.contains(d.getId())).collect(Collectors.toList()),
+				domainGroupCodeName);
+	}
+
+	@PatchMapping("/group/{domainGroupId}")
+	@Transactional
+	@PreAuthorize("hasRole('ROLE_SYSTEM_ADMIN')")
+	public DomainGroupView deleteDomainFromGroup(@PathVariable Long domainGroupId, @RequestBody Long domainId) {
+		return domainGroupService.deleteDomainFromGroup(
+				domainService.findDomain(domainId).orElseThrow(() -> new IllegalArgumentException(String.format("Domain with id %s doesn't exist", domainId))),
+				domainGroupId);
+	}
+
+	@PutMapping("/group/{domainGroupId}")
+	@Transactional
+	@PreAuthorize("hasRole('ROLE_SYSTEM_ADMIN')")
+	public Id updateDomainGroup(@PathVariable Long domainGroupId, @RequestBody DomainGroupView domainGroupView) {
+		 return new Id(domainGroupService.updateDomainGroup(domainGroupId, domainGroupView).getId());
 	}
 
 }
